requirements.txt
fastapi
uvicorn
pydantic
pdf2image
easyocr
opencv-python-headless
Pillow
diskcache
torch
torchvision
